+++ test/cypress/support/commands.js
@@ -35,3 +35,9 @@
 //     }
 //   }
 // }
+
+Cypress.on('uncaught:exception', (err, runnable) => {
+    // returning false here prevents Cypress from
+    // failing the test
+    return false
+})
